Fixed tests-build
Had some problems with building the tests with a combination of recent g++ and gtest versions.
`ASSERT_TRUE` on a `shared_ptr`/`optional` object seems to require explicit cast to `bool`.

Switched to `static_cast<bool>` instead of c-style casts. Seems to be related to more recent boost library that now uses explicit bool operators. Gtest seems to have a problem with that (google/googletest#429).